Model_train_microservice/requirements.txt
fastapi
uvicorn
tensorflow
pandas
numpy
scikit-learn
python-multipart
tensorflow-addons
pymongo
Pillow
